Update dependency: xhr
This version of xhr now passes `null` instead of `undefined` as a default body value to the `send` method of an instance of XMLHttpRequest. In most browsers, `undefined` had no ill-consequences when making a request, but as I discovered, Microsoft Edge was including `undefined` as the body of the request, which some servers don't like. `null` is a long-accepted way of indicating there is no body after the headers are sent.